Refactor capsule button into controls package
