+++ gui/controls/__init__.py
@@ -1,0 +1,1 @@
+"""Custom control widgets for the AutoML GUI."""
